# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tiles Stock Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads:
    get:
      operationId: listUploads
      tags: [uploads]
      summary: List recent PDF uploads
      parameters:
        - name: depotId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of uploads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Upload"
    post:
      operationId: uploadPdf
      tags: [uploads]
      summary: Upload a PDF stock report for a depot
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "201":
          description: Upload created and processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /uploads/{id}:
    get:
      operationId: getUpload
      tags: [uploads]
      summary: Get upload status and details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Upload details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Upload:
      type: object
      properties:
        id:
          type: integer
        depotId:
          type: integer
        depotName:
          type: string
        filename:
          type: string
        status:
          type: string
          enum: [pending, processing, done, failed]
        itemsExtracted:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        stockDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - depotId
        - depotName
        - filename
        - status
        - createdAt
    UploadInput:
      type: object
      properties:
        depotId:
          type: integer
        stockDate:
          type: string
          format: date
        file:
          type: string
          format: binary
      required:
        - depotId
        - file
...